Add Hashbase
Hashbase is a public peer service for Dat archives. It provides a HTTP-accessible interface for creating an account and uploading Dats. It was created to power a content-community for the Beaker Browser. The license for the source code is available at https://github.com/beakerbrowser/hashbase/blob/master/LICENSE and a demo website made with TiddlyWiki and hosted with Hashbase is available at https://tromsitegr.hashbase.io/#:HOME .